# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Asist Employed API
  description: |
    API REST del sistema de control de asistencia de empleados.

    - Autenticación con JWT (header `Authorization: Bearer <token>`).
    - Todos los endpoints requieren token salvo `POST /auth/login` y `GET /health`.
    - Los timestamps los fija el servidor; el cliente nunca envía horas de fichaje.
    - Las contraseñas se almacenan con bcrypt y nunca se devuelven en ninguna respuesta.
  version: 1.0.0
  license:
    name: MIT
servers:
  - url: /api/v1
...
paths:
...
  /employees:
    get:
      tags: [employees]
      summary: Listar empleados
      description: Solo admin. Devuelve usuarios con rol `employee`, sin contraseñas.
      operationId: listEmployees
      responses:
        '200':
          description: Lista de empleados
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [employees]
      summary: Crear empleado
      description: Solo admin. La contraseña se hashea con bcrypt antes de guardarse.
      operationId: createEmployee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEmployeeRequest'
      responses:
        '201':
          description: Empleado creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: Ya existe un usuario con ese email
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /employees/{id}:
    parameters:
      - $ref: '#/components/parameters/IdParam'
    get:
      tags: [employees]
      summary: Obtener un empleado
      description: Admin puede ver a cualquiera; un empleado solo a sí mismo.
      operationId: getEmployee
      responses:
        '200':
          description: Empleado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [employees]
      summary: Actualizar empleado
      description: Solo admin. Si se envía `password`, se re-hashea.
      operationId: updateEmployee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEmployeeRequest'
      responses:
        '200':
          description: Empleado actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Ya existe un usuario con ese email
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags: [employees]
      summary: Eliminar empleado
      description: Solo admin.
      operationId: deleteEmployee
      responses:
        '204':
          description: Empleado eliminado
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    Unauthorized:
      description: Token ausente, inválido o expirado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: El rol del usuario no permite esta operación
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ValidationError:
      description: Cuerpo o parámetros inválidos
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Código estable para manejo programático
              examples: [INVALID_CREDENTIALS, VALIDATION_ERROR, FORBIDDEN, NOT_FOUND, CONFLICT]
            message:
              type: string
              description: Descripción legible del error
...
    User:
      type: object
      required: [id, email, name, role, department, position, created_at]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        role:
          type: string
          enum: [employee, admin]
        department:
          type: string
        position:
          type: string
        created_at:
          type: string
          format: date-time
    CreateEmployeeRequest:
      type: object
      required: [name, email, password, department, position]
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
          description: Mínimo 8 caracteres
        department:
          type: string
          minLength: 1
        position:
          type: string
          minLength: 1
    UpdateEmployeeRequest:
      type: object
      minProperties: 1
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        department:
          type: string
          minLength: 1
        position:
          type: string
          minLength: 1
...
security:
  - bearerAuth: []